requirements.txt
  
Django
django-ckeditor
django-cleanup
django-crispy-forms
django-js-asset